ml_service/requirements.txt
flask
pandas
numpy
scikit-learn
joblibpip 